Go implementation of walking rationals.
The implementation is more optimized than the Python version,
by carefully choosing a data structure so that all operations
are cheap.

The 2x2 matrix associated to a continued fraction is kept
in the structure so that switching to a neighbouring fraction
only requires minimal recomputation.

Using this structure, more than 1e8 values can be enumerated
per second.